+++ utils/transfer-to-mongodb.py
@@ -3,7 +3,7 @@ from sklearn.metrics.pairwise import cosine_similarity
 import numpy as np
 from pymongo import MongoClient
 import datetime
-import cv2  # OpenCV for webcam access
+import cv2 
 
 # Connect to MongoDB
 client = MongoClient("mongodb://Krrish:krrish123@ac-7gvopvc-shard-00-00.1nxbntm.mongodb.net:27017,ac-7gvopvc-shard-00-01.1nxbntm.mongodb.net:27017,ac-7gvopvc-shard-00-02.1nxbntm.mongodb.net:27017/sih?replicaSet=atlas-11fue8-shard-0&ssl=true&authSource=admin&w=majority&appName=Cluster1")  # Change to your MongoDB URI if needed
